refactor: Change `renderer.jsx` to `renderer.tsx`
Refactor to use TypeScript in renderer.tsx, update relevant and outdated docs.
