+++ package.json
@@ -1,5 +1,7 @@
 {
     "dependencies": {
+        "@types/react": "^18.3.3",
+        "@types/react-dom": "^18.3.0",
         "dompurify": "^3.1.5",
         "highlight.js": "^11.9.0",
         "immer": "^10.1.1",
@@ -22,6 +24,8 @@
         "@babel/preset-env": "^7.24.7",
         "@babel/preset-react": "^7.24.7",
         "@babel/preset-typescript": "^7.24.7",
+        "@types/markdown-it": "^14.1.1",
+        "@types/throttle-debounce": "^5.0.2",
         "babel-loader": "^9.1.3",
         "webpack": "^5.92.1",
         "webpack-cli": "^5.1.4",
