+++ tsconfig.json
@@ -12,7 +12,7 @@
         "allowJs": true,
         "target": "ES2021",
         "module": "NodeNext",
-        "moduleResolution": "NodeNext"
+        "moduleResolution": "NodeNext",
     },
     "exclude": [
         "./dist/**/*" // This is what fixed it!
